fixed dependency bug somehow
It kept throwing an error that it couldn't install dependencies (specifically ggplot2) which was weird bc it was already installed. I restarted my R studio session and tried document() one last time and it worked. Not sure what the bug is there, maybe it has something to do with installing over the package when an older version is loaded already.